+++ backend/routes/playlists.js
@@ -2,23 +2,11 @@ const express = require('express');
 const router = express.Router();
 const playlistService = require('../services/playlistService');
 
-// POST import a playlist from external music service
+// POST import playlists from external music service
 router.post('/import', async (req, res) => {
     try {
-        const { userId, playlistData } = req.body;
-        const result = await playlistService.importPlaylist(userId, playlistData);
-        res.json(result);
-    } catch (error) {
-        res.status(500).send(error.message);
-    }
-});
-
-// GET export a playlist to external music service
-router.get('/export/:playlistId', async (req, res) => {
-    try {
-        const { playlistId } = req.params;
-        const { platform } = req.query;
-        const result = await playlistService.exportPlaylist(playlistId, platform);
+        const { userId, externalService } = req.body;
+        const result = await playlistService.importPlaylists(userId, externalService);
         res.json(result);
     } catch (error) {
         res.status(500).send(error.message);
