--- backend/services/musicService.js
@@ -1,14 +1,0 @@
-// musicnService.js
-
-const searchMusic = async (query) => {
-    // Logic to perform a search across connected music platforms
-};
-
-const getSongDetails = async (songId) => {
-    // Logic to retrieve detailed information about a song from connected platforms
-};
-
-module.exports = {
-    searchMusic,
-    getSongDetails,
-};
