+++ backend/services/searchService.js
@@ -1,0 +1,14 @@
+// searchService.js
+
+const searchMusic = async (query) => {
+    // Logic to perform a search
+};
+
+const getSongDetails = async (songId) => {
+    // Logic to retrieve detailed information about a song
+};
+
+module.exports = {
+    searchMusic,
+    getSongDetails,
+};
